+++ codecs/CHANGELOG.md
@@ -8,6 +8,10 @@ SPDX-License-Identifier: GPL-3.0-or-later
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Add Nostr NIP-19 encoding support.
+
 ## [0.0.0]
 
 - Reserve crate name.